Add cluster, type query params to type/sector listing endpoints
